moved fetch calls to child components
