+++ src/screens/shared/dashboard/StatCards.js
@@ -1,4 +1,6 @@
 import React from 'react';
+
+import { Chat, WbSunny, Check, Assignment } from '@material-ui/icons';
 import { withStyles } from '@material-ui/core/styles';
 
 const styles = {
@@ -21,8 +23,49 @@ const styles = {
   }
 };
 class StatCards extends React.PureComponent {
+  getCardStats = () => {
+    // TODO: Replace with Airtable Call
+    const iconProps = {
+      fontSize: 'large',
+      style: { color: 'var(--ks-dark-blue)' }
+    };
+
+    return [
+      {
+        icon: <Chat {...iconProps} />,
+        name: 'Total Harvest',
+        number: '17,000',
+        unit: ' lbs',
+        description: 'of harvest to date'
+      },
+      {
+        icon: <Check {...iconProps} />,
+        name: 'GAP Certification',
+        number: '40',
+        unit: '%',
+        description: 'of farms in Food Hub are GAP Certified'
+      },
+      {
+        icon: <WbSunny {...iconProps} />,
+        name: 'Harvesting Farms',
+        number: '10',
+        unit: ' farms',
+        description: 'are harvesting this week'
+      },
+      {
+        icon: <Assignment {...iconProps} />,
+        name: 'North Shore GAP',
+        number: '20',
+        unit: '%',
+        description: 'of all GAP-certified farms in North Shore are NSEVP'
+      }
+    ];
+  };
+
   render() {
-    const { classes, stats } = this.props;
+    const { classes } = this.props;
+    const stats = this.getCardStats();
+
     return (
       <div className={classes.root}>
         {stats.map(stat => (
